requirements.txt
python-dotenv
langchain
langchainhub
langchain-openai
langchain-community
langchain-ollama
flask